test(ci): split resolver workflow contract
