+++ test/pr-workflow-contract.test.ts
@@ -13,7 +13,6 @@ import {
   type WorkflowJob,
   type WorkflowStep,
 } from "./helpers/e2e-workflow-contract";
-import { execTimeout } from "./helpers/timeouts";
 
 type CiWorkflow = {
   "run-name"?: string;
@@ -246,9 +245,6 @@ describe("pull request and main workflow contracts", () => {
       ".github/actions/ci-installer-integration/action.yaml",
     ),
   };
-  const resolveHermesBaseAction = readYaml<CompositeAction>(
-    ".github/actions/resolve-hermes-base-image/action.yaml",
-  );
 
   // source-shape-contract: security -- Installer hashes must be verified by base-trusted or immutable bootstrap code
   it("runs pull request installer verification from immutable trusted code", () => {
@@ -1203,108 +1199,6 @@ describe("pull request and main workflow contracts", () => {
     expect(mainSuccess.status).toBe(0);
     expect(mainFailure.status).not.toBe(0);
     expect(mainFailure.stdout).toContain("real-openclaw-dist-harness failed");
-  });
-
-  it("rejects a pulled Hermes base without MCP HTTP imports and falls back locally", () => {
-    const temp = mkdtempSync(join(tmpdir(), "nemoclaw-hermes-base-resolver-"));
-    const fakeBin = join(temp, "bin");
-    const dockerLog = join(temp, "docker.log");
-    const githubEnv = join(temp, "github.env");
-    const remoteDigest = `ghcr.io/nvidia/nemoclaw/hermes-sandbox-base@sha256:${"a".repeat(64)}`;
-    const resolver = requiredStep(resolveHermesBaseAction, "Resolve Hermes sandbox base image").run;
-
-    try {
-      mkdirSync(fakeBin);
-      writeFileSync(githubEnv, "");
-      writeFileSync(
-        join(fakeBin, "docker"),
-        [
-          "#!/usr/bin/env bash",
-          "set -u",
-          'printf "%s\\0" "$@" >> "$DOCKER_LOG"',
-          'printf "\\0" >> "$DOCKER_LOG"',
-          'if [[ "$1" == "pull" || "$1" == "build" ]]; then exit 0; fi',
-          'if [[ "$1" == "image" && "$2" == "inspect" ]]; then',
-          '  printf "%s\\n" "$REMOTE_DIGEST"',
-          "  exit 0",
-          "fi",
-          'if [[ "$1" == "run" ]]; then',
-          '  entrypoint=""',
-          '  image=""',
-          "  while (($#)); do",
-          '    if [[ "$1" == "--entrypoint" ]]; then',
-          '      entrypoint="$2"',
-          '      image="$3"',
-          "      break",
-          "    fi",
-          "    shift",
-          "  done",
-          '  if [[ "$entrypoint" == "/usr/bin/ldd" ]]; then',
-          '    printf "ldd (Ubuntu GLIBC 2.39) 2.39\\n"',
-          "    exit 0",
-          "  fi",
-          '  if [[ "$entrypoint" == "sh" ]]; then exit 0; fi',
-          '  if [[ "$entrypoint" == "/opt/hermes/.venv/bin/python" ]]; then',
-          '    [[ "$image" != "$REMOTE_DIGEST" ]]',
-          "    exit",
-          "  fi",
-          "fi",
-          'printf "unexpected docker invocation:" >&2',
-          'printf " %q" "$@" >&2',
-          'printf "\\n" >&2',
-          "exit 2",
-          "",
-        ].join("\n"),
-        { mode: 0o755 },
-      );
-      // Keep the fake executable in a dedicated PATH directory so every other
-      // command in the composite action remains the real host utility.
-      const result = spawnSync("bash", ["-c", resolver ?? ""], {
-        cwd: process.cwd(),
-        encoding: "utf8",
-        timeout: execTimeout(),
-        env: {
-          ...process.env,
-          DOCKER_LOG: dockerLog,
-          GITHUB_ACTION_PATH: join(process.cwd(), ".github/actions/resolve-hermes-base-image"),
-          GITHUB_ENV: githubEnv,
-          GITHUB_SHA: "1".repeat(40),
-          PATH: `${fakeBin}:${process.env.PATH ?? ""}`,
-          REMOTE_DIGEST: remoteDigest,
-        },
-      });
-
-      expect(result.status, result.stderr).toBe(0);
-      expect(result.stdout).toContain("lacks the packaged MCP Streamable HTTP client imports");
-      expect(result.stdout).toContain("building locally");
-      expect(readFileSync(githubEnv, "utf8").trim()).toBe(
-        "HERMES_BASE_IMAGE=nemoclaw-hermes-base-local",
-      );
-
-      const calls = readFileSync(dockerLog, "utf8")
-        .split("\0\0")
-        .filter(Boolean)
-        .map((call) => call.split("\0").filter(Boolean));
-      const firstPull = calls.find((args) => args[0] === "pull");
-      expect(firstPull?.[0]).toBe("pull");
-      expect(firstPull?.[1]).toMatch(
-        /^ghcr\.io\/nvidia\/nemoclaw\/hermes-sandbox-base@sha256:[0-9a-f]{64}$/,
-      );
-      const remoteProbe = calls.findIndex(
-        (args) => args.includes("/opt/hermes/.venv/bin/python") && args.includes(remoteDigest),
-      );
-      const localBuild = calls.findIndex((args) => args[0] === "build");
-      const localProbe = calls.findIndex(
-        (args) =>
-          args.includes("/opt/hermes/.venv/bin/python") &&
-          args.includes("nemoclaw-hermes-base-local"),
-      );
-      expect(remoteProbe).toBeGreaterThanOrEqual(0);
-      expect(localBuild).toBeGreaterThan(remoteProbe);
-      expect(localProbe).toBeGreaterThan(localBuild);
-    } finally {
-      rmSync(temp, { force: true, recursive: true });
-    }
   });
 
   // source-shape-contract: security -- CI dependency installs must never execute package lifecycle scripts from fetched code
